fix(rtc): reconcile retained peer expiry
