+++ packages/tests/shared/webrtc-group-manager.test.ts
@@ -7,40 +7,18 @@ import {
 
 import type { ClientInfo, OverlayInfo } from '@shared/api/api-config.ts';
 import { toScopedOverlayId } from '@shared/api/api-type-utils.ts';
-// dprint-ignore
-import { createDefaultGroupLifecyclePolicy } from '@shared/api/group-lifecycle/group-lifecycle-policy-presets.ts';
-import { toGroupMemberPolicy } from '@shared/api/group-lifecycle/to-normalized-group-lifecycle-policy.ts';
-import type {
-    AuditStamp,
-    Group,
-    GroupMember,
-    GroupPresenceSession,
-    GroupSnapshot
-} from '@shared/api/group-types.ts';
+import type { GroupSnapshot } from '@shared/api/group-types.ts';
 import { LatestRepository } from '@shared/cache/LatestRepository.ts';
-import type { WebRtcConnectionService } from '@shared/services/web-rtc-connection-service.ts';
 import { WebRtcGroupManager } from '@shared/services/web-rtc-group-manager.ts';
 import type { WebRtcGroupPeerSelection } from '@shared/services/webrtc-group-manager-contracts.ts';
-import { createTestGroup } from '../create-test-group.ts';
-import type { SimulatedNativeRtcPeerConnection } from './native-rtc-connection-fixture.ts';
 import { createSimulatedRtcConnections } from './simulated-rtc-connection-service.ts';
-
-interface GroupSnapshotFixture {
-    readonly groupId: string;
-    readonly membershipVersion: number;
-    readonly memberSessionIds: readonly string[];
-    readonly applicationId?: string;
-    readonly workspaceId?: string;
-    readonly maxConcurrentEdgeSetups?: number;
-}
-
-interface RtcConnectionHarness {
-    readonly service: WebRtcConnectionService;
-    knownPeerIds(): string[];
-    peerIdsWithNoReconnectableLanes(): string[];
-    markReconnectable(peerId: string): boolean;
-    nativePeer(peerId: string): SimulatedNativeRtcPeerConnection;
-}
+import {
+    acceptActiveLayoutGroup,
+    createClientInfo,
+    createGroupSnapshot,
+    createOverlayInfo,
+    createRtcConnectionHarness
+} from './web-rtc-group-manager-test-fixture.ts';
 
 describe('WebRtcGroupManager', () => {
     it('notifies with final desired and canonical degree-limited RTT peers after dial reconciliation', async () => {
@@ -917,217 +895,6 @@ describe('WebRtcGroupManager', () => {
         expect(manager.readDiagnostics().connectDeferredBudgetCount).toBe(0);
     });
 
-    it('bounds in-flight setups per group and dials paced peers as setups complete', async () => {
-        const groupCache = new LatestRepository<string, GroupSnapshot>();
-        const clientCache = new LatestRepository<string, ClientInfo>();
-        const rtcQBox = createRtcConnectionHarness('self');
-        const acceptedOverlayCache = new LatestRepository<string, OverlayInfo>();
-        const manager = new WebRtcGroupManager(
-            rtcQBox.service,
-            { groupCache, clientCache, acceptedOverlayCache },
-            { overlayTransitionGraceMs: 0 }
-        );
-        manager.startReconcileWakes();
-        const peerIds = ['peer-a', 'peer-b', 'peer-c', 'peer-d', 'peer-e'];
-        for (const peerId of peerIds) {
-            clientCache.set(peerId, createClientInfo(peerId, true));
-        }
-
-        await acceptActiveLayoutGroup(
-            manager,
-            acceptedOverlayCache,
-            createGroupSnapshot({
-                groupId: 'group-1',
-                membershipVersion: 1,
-                memberSessionIds: ['self', ...peerIds],
-                maxConcurrentEdgeSetups: 2
-            })
-        );
-
-        expect(rtcQBox.knownPeerIds()).toEqual(['peer-a', 'peer-b']);
-        expect(rtcQBox.service.inFlightPeerIds()).toEqual(['peer-a', 'peer-b']);
-        expect(manager.readDiagnostics()).toMatchObject({ connectAttemptCount: 2, connectDeferredPacingCount: 3 });
-
-        rtcQBox.nativePeer('peer-a').setConnected();
-        await manager.whenReconciled();
-
-        expect(rtcQBox.knownPeerIds()).toEqual(['peer-a', 'peer-b', 'peer-c']);
-        expect(rtcQBox.service.inFlightPeerIds()).toEqual(['peer-b', 'peer-c']);
-
-        rtcQBox.nativePeer('peer-b').setConnected();
-        rtcQBox.nativePeer('peer-c').setConnected();
-        await manager.whenReconciled();
-
-        expect(rtcQBox.knownPeerIds()).toEqual(['peer-a', 'peer-b', 'peer-c', 'peer-d', 'peer-e']);
-        expect(manager.readDiagnostics()).toMatchObject({ connectAttemptCount: 5 });
-    });
-
-    it('wakes a reconcile when a setup ends by removal while a paced dial is waiting', async () => {
-        const groupCache = new LatestRepository<string, GroupSnapshot>();
-        const clientCache = new LatestRepository<string, ClientInfo>();
-        const rtcQBox = createRtcConnectionHarness('self');
-        const acceptedOverlayCache = new LatestRepository<string, OverlayInfo>();
-        const manager = new WebRtcGroupManager(
-            rtcQBox.service,
-            { groupCache, clientCache, acceptedOverlayCache },
-            { overlayTransitionGraceMs: 0 }
-        );
-        manager.startReconcileWakes();
-        for (const peerId of ['peer-a', 'peer-b']) {
-            clientCache.set(peerId, createClientInfo(peerId, true));
-        }
-        await acceptActiveLayoutGroup(
-            manager,
-            acceptedOverlayCache,
-            createGroupSnapshot({
-                groupId: 'group-1',
-                membershipVersion: 1,
-                memberSessionIds: ['self', 'peer-a', 'peer-b'],
-                maxConcurrentEdgeSetups: 1
-            })
-        );
-        expect(manager.readDiagnostics()).toMatchObject({ reconcileRunCount: 1, connectAttemptCount: 1 });
-        const lifecycle: string[] = [];
-        rtcQBox.service.onRtcPeerLifecycleDo('observer-after-manager', {
-            onCreated: (peer) => {
-                lifecycle.push(`created:${peer.peerId}`);
-            },
-            onDeleted: (peer) => {
-                lifecycle.push(`deleted:${peer.peerId}`);
-            }
-        });
-
-        rtcQBox.service.disconnectPeer('peer-a');
-        expect(lifecycle).toEqual(['deleted:peer-a']);
-        await manager.whenReconciled();
-
-        // The wake ran after the deletion notice completed, so a later observer
-        // never sees the re-dial before the ending that released it.
-        expect(lifecycle).toEqual(['deleted:peer-a', 'created:peer-a']);
-        expect(rtcQBox.knownPeerIds()).toEqual(['peer-a']);
-        expect(manager.readDiagnostics()).toMatchObject({ reconcileRunCount: 2, connectAttemptCount: 2 });
-        manager.stopReconcileWakes();
-    });
-
-    it('retries a desired peer when its setup closes without a paced dial waiting', async () => {
-        const groupCache = new LatestRepository<string, GroupSnapshot>();
-        const clientCache = new LatestRepository<string, ClientInfo>();
-        const rtcQBox = createRtcConnectionHarness('self');
-        const acceptedOverlayCache = new LatestRepository<string, OverlayInfo>();
-        const manager = new WebRtcGroupManager(
-            rtcQBox.service,
-            { groupCache, clientCache, acceptedOverlayCache },
-            { overlayTransitionGraceMs: 0 }
-        );
-        manager.startReconcileWakes();
-        clientCache.set('peer-a', createClientInfo('peer-a', true));
-        await acceptActiveLayoutGroup(
-            manager,
-            acceptedOverlayCache,
-            createGroupSnapshot({
-                groupId: 'group-1',
-                membershipVersion: 1,
-                memberSessionIds: ['self', 'peer-a']
-            })
-        );
-        expect(manager.readDiagnostics()).toMatchObject({
-            reconcileRunCount: 1,
-            connectAttemptCount: 1,
-            connectDeferredPacingCount: 0
-        });
-
-        rtcQBox.nativePeer('peer-a').close();
-        await manager.whenReconciled();
-
-        expect(rtcQBox.knownPeerIds()).toEqual(['peer-a']);
-        expect(manager.readDiagnostics()).toMatchObject({
-            reconcileRunCount: 2,
-            connectAttemptCount: 2
-        });
-        manager.stopReconcileWakes();
-    });
-
-    it('stands down a wake scheduled before the wakes stop', async () => {
-        const groupCache = new LatestRepository<string, GroupSnapshot>();
-        const clientCache = new LatestRepository<string, ClientInfo>();
-        const rtcQBox = createRtcConnectionHarness('self');
-        const acceptedOverlayCache = new LatestRepository<string, OverlayInfo>();
-        const manager = new WebRtcGroupManager(
-            rtcQBox.service,
-            { groupCache, clientCache, acceptedOverlayCache },
-            { overlayTransitionGraceMs: 0 }
-        );
-        manager.startReconcileWakes();
-        for (const peerId of ['peer-a', 'peer-b']) {
-            clientCache.set(peerId, createClientInfo(peerId, true));
-        }
-        await acceptActiveLayoutGroup(
-            manager,
-            acceptedOverlayCache,
-            createGroupSnapshot({
-                groupId: 'group-1',
-                membershipVersion: 1,
-                memberSessionIds: ['self', 'peer-a', 'peer-b'],
-                maxConcurrentEdgeSetups: 1
-            })
-        );
-        expect(rtcQBox.knownPeerIds()).toEqual(['peer-a']);
-
-        // The removal schedules the wake; shutdown stops the wakes before it runs.
-        rtcQBox.service.disconnectPeer('peer-a');
-        manager.stopReconcileWakes();
-        await manager.whenReconciled();
-
-        expect(rtcQBox.knownPeerIds()).toEqual([]);
-        expect(manager.readDiagnostics()).toMatchObject({ reconcileRunCount: 1 });
-    });
-
-    it('does not re-run a pass for an ending its own dial caused', async () => {
-        const groupCache = new LatestRepository<string, GroupSnapshot>();
-        const clientCache = new LatestRepository<string, ClientInfo>();
-        // peer-a's lane refuses on every dial, which resets and removes its native
-        // connection inside the pass; with peer-c paced, a latch on that removal
-        // would re-run the pass forever.
-        const rtcQBox = createRtcConnectionHarness('self', [], (peerId) => {
-            if (peerId === 'peer-a') {
-                throw new Error('lane refused');
-            }
-            return true;
-        });
-        const acceptedOverlayCache = new LatestRepository<string, OverlayInfo>();
-        const manager = new WebRtcGroupManager(
-            rtcQBox.service,
-            { groupCache, clientCache, acceptedOverlayCache },
-            { overlayTransitionGraceMs: 0 }
-        );
-        manager.startReconcileWakes();
-        for (const peerId of ['peer-a', 'peer-b', 'peer-c']) {
-            clientCache.set(peerId, createClientInfo(peerId, true));
-        }
-        await acceptActiveLayoutGroup(
-            manager,
-            acceptedOverlayCache,
-            createGroupSnapshot({
-                groupId: 'group-1',
-                membershipVersion: 1,
-                memberSessionIds: ['self', 'peer-a', 'peer-b', 'peer-c'],
-                maxConcurrentEdgeSetups: 1
-            })
-        );
-        expect(rtcQBox.knownPeerIds()).toEqual(['peer-b']);
-
-        await manager.ensureAllGroupsConnected();
-        await manager.whenReconciled();
-
-        // Pass two paces both peer-a and peer-c behind peer-b's setup.
-        expect(manager.readDiagnostics()).toMatchObject({
-            reconcileRunCount: 2,
-            reconcileCoalescedRerunCount: 0,
-            connectDeferredPacingCount: 3
-        });
-        manager.stopReconcileWakes();
-    });
-
     it('frees a pacing slot at once when an admitted dial starts nothing', async () => {
         const groupCache = new LatestRepository<string, GroupSnapshot>();
         const clientCache = new LatestRepository<string, ClientInfo>();
@@ -1229,84 +996,6 @@ describe('WebRtcGroupManager', () => {
         expect(manager.readDiagnostics()).toMatchObject({ connectDeferredBudgetCount: 0 });
     });
 
-    it('does not wake a reconcile for a setup ending while no dial is waiting, nor after wakes stop', async () => {
-        const groupCache = new LatestRepository<string, GroupSnapshot>();
-        const clientCache = new LatestRepository<string, ClientInfo>();
-        const rtcQBox = createRtcConnectionHarness('self');
-        const acceptedOverlayCache = new LatestRepository<string, OverlayInfo>();
-        const manager = new WebRtcGroupManager(
-            rtcQBox.service,
-            { groupCache, clientCache, acceptedOverlayCache },
-            { overlayTransitionGraceMs: 0 }
-        );
-        manager.startReconcileWakes();
-        for (const peerId of ['peer-a', 'peer-b']) {
-            clientCache.set(peerId, createClientInfo(peerId, true));
-        }
-        await acceptActiveLayoutGroup(
-            manager,
-            acceptedOverlayCache,
-            createGroupSnapshot({ groupId: 'group-1', membershipVersion: 1, memberSessionIds: ['self', 'peer-a', 'peer-b'] })
-        );
-        expect(manager.readDiagnostics()).toMatchObject({ reconcileRunCount: 1, connectDeferredPacingCount: 0 });
-
-        rtcQBox.nativePeer('peer-a').setConnected();
-        await manager.whenReconciled();
-        expect(manager.readDiagnostics().reconcileRunCount).toBe(1);
-
-        await manager.acceptGroupUpdate(
-            createGroupSnapshot({
-                groupId: 'group-1',
-                membershipVersion: 2,
-                memberSessionIds: ['self', 'peer-a', 'peer-b'],
-                maxConcurrentEdgeSetups: 1
-            })
-        );
-        manager.stopReconcileWakes();
-        rtcQBox.nativePeer('peer-b').setConnected();
-        await manager.whenReconciled();
-
-        expect(manager.readDiagnostics().reconcileRunCount).toBe(2);
-    });
-
-    it('charges a peer wanted by two groups to both and waits while either owner is at its bound', async () => {
-        const groupCache = new LatestRepository<string, GroupSnapshot>();
-        const clientCache = new LatestRepository<string, ClientInfo>();
-        const rtcQBox = createRtcConnectionHarness('self');
-        const acceptedOverlayCache = new LatestRepository<string, OverlayInfo>();
-        const manager = new WebRtcGroupManager(
-            rtcQBox.service,
-            { groupCache, clientCache, acceptedOverlayCache },
-            { overlayTransitionGraceMs: 0 }
-        );
-        manager.startReconcileWakes();
-        for (const peerId of ['peer-a', 'peer-shared', 'peer-b']) {
-            clientCache.set(peerId, createClientInfo(peerId, true));
-        }
-        const saturated = createGroupSnapshot({
-            groupId: 'group-saturated',
-            membershipVersion: 1,
-            memberSessionIds: ['self', 'peer-a', 'peer-shared'],
-            maxConcurrentEdgeSetups: 1
-        });
-        const idle = createGroupSnapshot({
-            groupId: 'group-idle',
-            membershipVersion: 1,
-            memberSessionIds: ['self', 'peer-shared', 'peer-b'],
-            maxConcurrentEdgeSetups: 5
-        });
-
-        await acceptActiveLayoutGroup(manager, acceptedOverlayCache, saturated);
-        await acceptActiveLayoutGroup(manager, acceptedOverlayCache, idle);
-
-        expect(rtcQBox.knownPeerIds()).toEqual(['peer-a', 'peer-b']);
-
-        rtcQBox.nativePeer('peer-a').setConnected();
-        await manager.whenReconciled();
-
-        expect(rtcQBox.knownPeerIds()).toEqual(['peer-a', 'peer-b', 'peer-shared']);
-    });
-
     it('counts connect failures in diagnostics', async () => {
         const groupCache = new LatestRepository<string, GroupSnapshot>();
         const clientCache = new LatestRepository<string, ClientInfo>();
@@ -1375,46 +1064,6 @@ interface ReconciledPeerSelectionObservation extends WebRtcGroupPeerSelection {
     readonly knownPeerIds: readonly string[];
 }
 
-async function acceptActiveLayoutGroup(
-    manager: WebRtcGroupManager,
-    acceptedOverlayCache: LatestRepository<string, OverlayInfo>,
-    group: GroupSnapshot
-): Promise<void> {
-    acceptedOverlayCache.set(
-        toScopedOverlayId(group.group),
-        createOverlayInfo(
-            group,
-            group.activeSessions
-                .map((session) => session.sessionId)
-                .filter((sessionId) => sessionId !== manager.rtcQBox.input.sessionId),
-            group.group.snapshotVersion
-        )
-    );
-    await manager.acceptGroupUpdate(group);
-}
-
-function createRtcConnectionHarness(
-    sessionId: string,
-    initiallyConnectedPeerIds: readonly string[] = [],
-    connect?: (peerId: string) => boolean
-): RtcConnectionHarness {
-    const simulation = createSimulatedRtcConnections(sessionId, connect);
-    const { service } = simulation;
-    for (const peerId of initiallyConnectedPeerIds) {
-        const result = service.ensurePeerConnectionStarted(peerId);
-        if (result.left) {
-            throw new Error(`Failed to seed simulated peer ${peerId}`);
-        }
-    }
-    return {
-        service,
-        knownPeerIds: () => [...service.knownPeerIds()],
-        peerIdsWithNoReconnectableLanes: () => [...service.peerIdsWithNoReconnectableLanes()],
-        markReconnectable: simulation.markReconnectable,
-        nativePeer: simulation.nativePeer
-    };
-}
-
 class CountingClientCache extends LatestRepository<string, ClientInfo> {
     keysCalls = 0;
 
@@ -1426,133 +1075,4 @@ class CountingClientCache extends LatestRepository<string, ClientInfo> {
     resetCounters(): void {
         this.keysCalls = 0;
     }
-}
-
-function createClientInfo(sessionId: string, isOnline: boolean): ClientInfo {
-    return {
-        clientId: sessionId,
-        sessionId,
-        isOnline
-    };
-}
-
-function createGroupSnapshot(
-    fixture: GroupSnapshotFixture
-): GroupSnapshot {
-    const applicationId = fixture.applicationId ?? 'app-1';
-    const workspaceId = fixture.workspaceId ?? 'workspace-1';
-    const ownerPrincipalId = fixture.memberSessionIds[0];
-    if (ownerPrincipalId === undefined) {
-        throw new Error('Group fixture requires an owner session');
-    }
-
-    const group = createTestGroup({
-        applicationId,
-        workspaceId,
-        groupId: fixture.groupId,
-        slug: fixture.groupId,
-        displayName: fixture.groupId,
-        activeMemberCount: fixture.memberSessionIds.length,
-        ownerPrincipalId,
-        snapshotVersion: fixture.membershipVersion,
-        metadataVersion: 0,
-        rosterVersion: fixture.membershipVersion,
-        presenceVersion: fixture.membershipVersion,
-        formationElectorate: [...fixture.memberSessionIds],
-        ...(fixture.maxConcurrentEdgeSetups === undefined
-            ? {}
-            : {
-                memberPolicy: {
-                    ...toGroupMemberPolicy(createDefaultGroupLifecyclePolicy()),
-                    maxConcurrentEdgeSetups: fixture.maxConcurrentEdgeSetups
-                }
-            }),
-        acceptedLayoutIdentity: {
-            groupRevision: fixture.membershipVersion,
-            presenceRevision: fixture.membershipVersion,
-            version: fixture.membershipVersion,
-            state: 'active'
-        },
-        created: createAuditStamp(1, ownerPrincipalId),
-        updated: createAuditStamp(fixture.membershipVersion, ownerPrincipalId)
-    });
-    return {
-        causalRevision: {
-            groupRevision: fixture.membershipVersion,
-            presenceRevision: fixture.membershipVersion
-        },
-        group,
-        members: fixture.memberSessionIds.map((sessionId) => createGroupMember(group, sessionId)),
-        activeSessions: fixture.memberSessionIds.map((sessionId) => createGroupPresenceSession(group, sessionId)),
-        memberCount: fixture.memberSessionIds.length,
-        onlineMemberCount: fixture.memberSessionIds.length
-    };
-}
-
-function createGroupMember(group: Group, sessionId: string): GroupMember {
-    return {
-        applicationId: group.applicationId,
-        workspaceId: group.workspaceId,
-        groupId: group.groupId,
-        principalId: sessionId,
-        role: sessionId === group.ownerPrincipalId ? 'owner' : 'member',
-        status: 'active',
-        joined: group.created,
-        updated: group.updated,
-        invitedByPrincipalId: null,
-        invitationExpiresAtEpochMs: null,
-        left: null,
-        removed: null,
-        banned: null
-    };
-}
-
-function createGroupPresenceSession(group: Group, sessionId: string): GroupPresenceSession {
-    return {
-        applicationId: group.applicationId,
-        workspaceId: group.workspaceId,
-        groupId: group.groupId,
-        sessionId,
-        principalId: sessionId,
-        generationId: `generation-${sessionId}`,
-        generationVersion: group.rosterVersion,
-        status: 'active',
-        connectedAtEpochMs: group.rosterVersion,
-        lastHeartbeatAtEpochMs: group.rosterVersion,
-        expiresAtEpochMs: group.rosterVersion + 60_000,
-        disconnectedAtEpochMs: null,
-        disconnectReason: null
-    };
-}
-
-function createAuditStamp(atEpochMs: number, principalId: string): AuditStamp {
-    return {
-        atEpochMs,
-        actor: { kind: 'principal', principalId },
-        reason: null,
-        traceId: null,
-        requestId: null
-    };
-}
-
-function createOverlayInfo(
-    group: GroupSnapshot,
-    nextHopSessionIds: readonly string[],
-    overlayVersion = 1
-): OverlayInfo {
-    return {
-        overlayId: toScopedOverlayId(group.group),
-        groupRef: group.group,
-        topology: 'tree',
-        sourceGroupStateCausalRevision: group.causalRevision,
-        state: 'active',
-        name: group.group.displayName,
-        createdByClientId: group.group.ownerPrincipalId,
-        createdAtEpochMs: group.group.created.atEpochMs,
-        nextHopSessionIds,
-        degreeLimit: Math.max(1, nextHopSessionIds.length),
-        overlayVersion,
-        updatedAtEpochMs: overlayVersion,
-        provenance: 'server'
-    };
 }
